feat: 이벤트 토픽명 변경, 1분마다 랭킹 갱신요청 이벤트 수신 처리

+++ apps/ranking-service/src/main/java/io/antcamp/rankingservice/domain/event/TotalAssetCalcuatedEvent.java
@@ -4,7 +4,8 @@ import java.util.List;
 import java.util.UUID;
 
 /**
- * 자산 서비스가 대회 종료 후 참가자별 최종 총자산을 계산한 뒤 발행하는 이벤트 payload. 랭킹 서비스가 컨슘하여 최종 순위를 Redis와 DB에 모두 저장한다.
+ * 자산 서비스가 대회 종료 후 참가자별 최종 총자산을 계산한 뒤 발행하는 이벤트.
+ * 랭킹 서비스가 컨슘하여 최종 순위를 Redis와 DB에 모두 저장한다.
  */
 public record TotalAssetCalcuatedEvent(
         UUID competitionId,
@@ -12,7 +13,7 @@ public record TotalAssetCalcuatedEvent(
 ) {
     public record ParticipantTotalAsset(
             UUID userId,
-            long totalAsset  // 단위: 원 (현금 + 보유주식 가치)
+            Double totalAsset  // 현금 + 보유주식 시가 합산 (소수점 포함)
     ) {
     }
 }
