+++ apps/ranking-service/src/main/java/io/antcamp/rankingservice/domain/repository/RankingRedisRepository.java
@@ -1,6 +1,5 @@
 package io.antcamp.rankingservice.domain.repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -9,7 +8,7 @@ import java.util.UUID;
  */
 public interface RankingRedisRepository {
     // 매매 체결시, 랭킹 갱신
-    void upsertScore(UUID competitionId, UUID userId, BigDecimal totalAsset);
+    void upsertScore(UUID competitionId, UUID userId, Double totalAsset);
 
     // 랭킹 단건 조회
     long getRank(UUID competitionId, UUID userId);   // 0-based
@@ -21,6 +20,6 @@ public interface RankingRedisRepository {
     List<RankingEntry> getTopRankings(UUID competitionId, long offset, long count);
 
     // 랭킹 목록 저장용 record
-    record RankingEntry(UUID userId, BigDecimal totalAsset, long rank) {
+    record RankingEntry(UUID userId, Double totalAsset, long rank) {
     }
 }
